requirements.txt
https://github.com/TelegramPlayGround/pyrogram/archive/unknown_errors.zip
tgcrypto
Pillow
motor
python-dotenv
pyromod
graphqlclient
APScheduler
pytz
aiohttp
six
asyncio
google-generativeai
deep-translator
colorama
html2text
google-api-python-client
google-auth-httplib2
google-auth-oauthlib